requirements.txt
flask>=3.0.0
flask-socketio>=5.3.0
flask-cors>=4.0.0
pyyaml>=6.0.1
gunicorn>=21.0.0
eventlet>=0.33.3
